requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.1.1
Werkzeug==2.3.7
gunicorn==21.2.0
python-dotenv==1.0.0
SQLAlchemy==2.0.36
WTForms==3.0.1
email-validator==2.0.0
cryptography==42.0.7
psycopg[binary]==3.2.9
flask-sock==0.7.0
gevent==24.2.1
# pywebpush==2.0.0  # Optional: for push notifications
requests>=2.31.0
pywebpush>=1.14.0
